% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGEhammer.R
\docType{package}
\name{PGEhammer}
\alias{PGEhammer}
\title{PGEhammer}
\description{
A series of utility functions that improve the experience of
  working with other software tools in Plasmodium genomic epidemiology. This
  includes functions for wrangling data and visualising outputs. Part of the
  \href{https://github.com/mrc-ide/PGEforge}{PGEforge} set of resources.
}
