% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangle_vcf2long.R
\name{vcf2long}
\alias{vcf2long}
\title{Convert vcf to long format}
\usage{
vcf2long(vcf)
}
\arguments{
\item{vcf}{object of class vcfR}
}
\description{
Convert a vcf into a long format data frame with sample ID, locus, alleles and read counts for each allele.
}
