% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pedmap.R
\name{vcf2pedmap}
\alias{vcf2pedmap}
\title{vcf to pedmap format}
\arguments{
\item{vcf_file}{file path to VCF file containing all variants (SNPs and indels)}

\item{gds_file}{file path to the output of the GDS file for SeqArray}

\item{ped_file}{file path to the output of ped file}

\item{map_file}{file path to the output of map file}
}
\description{
Function to convert processed vcf data into pedmap format ready to use for isoRelate
}
