% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangle_vcf2long.R
\name{vcf2long}
\alias{vcf2long}
\title{Square a vector of values}
\usage{
vcf2long(x)
}
\arguments{
\item{x}{object of class vcfR.}
}
\description{
Simple test function that demonstrates some of the features of
  this package by squaring an input vector of values.
}
\examples{
# Find square of first 100 values
square(1:100)
}
